backend/requirements.txt
fastapi
uvicorn
pillow 
python-dotenv
PyYAML
opencv-python-headless
termcolor
openai
google
google-genai
google-api-core
